blk-mq: add flag allowing drivers to request workqueue/process context
The conversion of the loop driver always offloads IO to a private
workqueue upon invocation of ->queue_rq(). Since blk-mq already
supports running from affinitized workqueues, we can let blk-mq do
this if the driver wishes for that to happen. This means the loop
conversion can be simplified - it just needs to set the appropriate
flag when registrering the tag set.

Signed-off-by: Jens Axboe <axboe@fb.com>